+++ src/init.c
@@ -8,14 +8,14 @@
  */
 
 /* .Call calls */
-extern SEXP pinbasic_eho(SEXP, SEXP, SEXP);
-extern SEXP pinbasic_linke(SEXP, SEXP, SEXP);
-extern SEXP pinbasic_simBS(SEXP, SEXP);
+extern SEXP _pinbasic_eho(SEXP, SEXP, SEXP);
+extern SEXP _pinbasic_linke(SEXP, SEXP, SEXP);
+extern SEXP _pinbasic_simulateBS(SEXP, SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
-  {"pinbasic_eho",   (DL_FUNC) &pinbasic_eho,   3},
-  {"pinbasic_linke", (DL_FUNC) &pinbasic_linke, 3},
-  {"pinbasic_simBS", (DL_FUNC) &pinbasic_simBS, 2},
+  {"_pinbasic_eho",   (DL_FUNC) &_pinbasic_eho,   3},
+  {"_pinbasic_linke", (DL_FUNC) &_pinbasic_linke, 3},
+  {"_pinbasic_simulateBS", (DL_FUNC) &_pinbasic_simulateBS, 2},
   {NULL, NULL, 0}
 };