+++ src/simulateBS.cpp
@@ -1,9 +1,22 @@
 #include <Rcpp.h>
 using namespace Rcpp;
 
+//' Simulate trading data
+//'
+//' Simulates a matrix consisting of synthetic data for daily buys and sells
+//'
+//' If names are not set for \code{param} or one or more of the vector names do not match the valid choices, they are internally set to
+//' \code{'alpha'}, \code{'delta'}, \code{'epsilon_b'}, \code{'epsilon_s'}, \code{'mu'} (in this order).
+//'
+//' @return \emph{numeric}: Matrix with \code{ndays} rows and two columns which are named \code{'Buys'} and \code{'Sells'}.
+//'
+//' @inheritParams pin_ll
+//' @param ndays \emph{integer}: Number of trading days for which aggregated buys and sells are simulated, defaults to 60
+//'
+//' @export simulateBS
 // [[Rcpp::export]]
 
-NumericMatrix simBS(NumericVector param, int ndays) {
+NumericMatrix simulateBS(NumericVector param, int ndays) {
   NumericMatrix res(ndays,2);
   IntegerVector states_ind = IntegerVector::create(0,1,2);
   NumericVector state_probs(3);